pyproject.toml
[project]
name = "friday-2"
version = "0.1.0"
description = "Friday 2.0 - Ecosysteme d'intelligence personnelle auto-heberge"
requires-python = ">=3.11"
license = {text = "UNLICENSED"}

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Database
    "asyncpg>=0.30.0",

    # Cache & Events
    "redis[hiredis]>=5.2.0",

    # HTTP client
    "httpx>=0.28.0",

    # Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # LLM - Claude Sonnet 4.5 (D17)
    "anthropic>=0.42.0",

    # NLP / RGPD
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
    "spacy>=3.8.0",

    # Vectorstore - pgvector dans PostgreSQL (D19)
    "pgvector>=0.3.6",

    # Telegram
    "python-telegram-bot>=21.0",

    # Workflow
    "langgraph>=0.2.45",

    # Logging
    "structlog>=24.4.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=9.0.0",
    "croniter>=5.0.0",
    "schedule>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "isort>=5.13.0",
    "flake8>=7.1.0",
    "mypy>=1.13.0",
    "sqlfluff>=3.2.0",
    "pre-commit>=4.0.0",
    "types-redis>=4.6.0",
]

stt = [
    "faster-whisper>=1.1.0",
]

tts = [
    "kokoro>=0.4.0",
]

ocr = [
    "surya-ocr>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents/src"]

# ---------------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py311"]

# ---------------------------------------------------------------------------
# isort
# ---------------------------------------------------------------------------
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["agents", "services", "config"]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "presidio_analyzer.*",
    "presidio_anonymizer.*",
    "pgvector.*",
    "langgraph.*",
    "spacy.*",
    "faster_whisper.*",
    "kokoro.*",
    "surya.*",
    "n8n.*",
]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require services)",
    "e2e: End-to-end tests (full stack)",
    "slow: Slow tests (>10s)",
]
addopts = "-v --tb=short"

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
source = ["agents/src"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70

# ---------------------------------------------------------------------------
# flake8 (via setup.cfg or .flake8 since flake8 doesn't read pyproject.toml)
# ---------------------------------------------------------------------------

# ---------------------------------------------------------------------------
# sqlfluff
# ---------------------------------------------------------------------------
[tool.sqlfluff.core]
dialect = "postgres"
max_line_length = 100
# Exclude purely stylistic rules that have pre-existing violations in legacy migrations.
# These rules do not affect SQL correctness or safety.
exclude_rules = "CP01,CP02,CP03,CP04,CP05,LT01,LT02,LT03,LT04,LT05,LT06,LT07,LT08,LT09,LT10,LT11,LT12,LT13,RF01,RF02,RF03,RF04,AL01,AL02,AL03,AL04,AL05,AM01,AM02,AM03,AM04,AM05,AM06,JJ01"
